#2 [FEAT] - 물품 보관 현황 디테일 조회(Request, Pending, Stored) API 수정
